+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2022-07-13
+
+### Added
+- added the option for the Eigen layer to support semi-definite predictions
+- new option `n_zero_eigvals` for Eigen layer to specify how many zero eigenvalues the model should expect
+
+### Changed
+- Eigen layer now uses `torch.linalg.eigh` instead of `torch.linalg.eig`
+
 ## [0.2.0] - 2022-06-23
 
 ### Changed